requirements.txt
pyttsx3==2.90
SpeechRecognition==3.8.1
Pillow==9.3.0
gTTS==2.3.1
pygame==2.5.0